+++ package.json
@@ -25,7 +25,7 @@
   ],
   "devDependencies": {
     "concurrently": "^7.6.0",
-    "eslint": "^8.31.0",
+    "eslint": "^8.2.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-jsx-a11y": "^6.5.1",
@@ -40,9 +40,5 @@
       "npm run prettier --workspaces --if-present",
       "npm run lint-dry-run --workspaces --if-present"
     ]
-  },
-  "dependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/plugin-proposal-private-methods": "^7.18.6"
   }
 }
