+++ server/package.json
@@ -29,6 +29,7 @@
   "homepage": "https://github.com/Banyan-Labs/MoneyManageApp#readme",
   "description": "",
   "devDependencies": {
+    "dotenv": "^16.0.3",
     "eslint": "^8.31.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-base": "^15.0.0",
