--- server/routes/user.routes.js
@@ -1,11 +1,0 @@
-const { default: App } = require("../../money-managed-app/Client/src/App");
-
-const Controller = require('../controllers/user.controller');
-
-
-
-
-module.exports = app => {
-    app.post("/api/register", Controller.checkDuplicateEmail)
-   
-}
